Wind Visualizations
This R Markdown file contains three visualizations I made with the wind data we had. The map data isn't as great as I would like, but I'll try to work on making it prettier sometime next week. The grey areas on the map are states we don't have data for.